Added a new platform sidecar annotation
This new annotation prefix will be used to indicate to the executor
which containers are to be treated like platform sidecars (indicating
they should be launched before user workloads, and perhaps have different
ordering and failure modes).

Annotation are per-pod, so this is a "prefix". Individual *containers*
would add an annotation to the pod like this:

```
annotations:
  type.container.netflix.com/envoy: platform
containers:
  - name: envoy
  ...
```