Added image tag annotation for v1 schema pods
We need to know the original Tag that was used for container
images for security purposes.

This new annotation allows us to set and get that original tag
on the pod, so that later when we go to verify that pods sig
against what was originally submitted to titus (the tag), we
can verify.