Added new InjectedEnvVarName pod config for metatron TITUS-5821
For identity purposes, we need a way to distinguish between user-defined
env vars, titus-defined env vars, and the new *injected* env vars.

These injected env vars come from the new platform sidecar mutation
code, which has the power to inject env variables into the *users* containers
(not just the sidecars).

This conflicts with metatron verification, because now it sees this
extra var from nowhere!

This adds a new annotation so the mutator and add to it, and then
the titus-executor and ignore these vars.

This change kinda assumes (safely for now) that the list of vars
is evenly applied to all user containers (usually just the main one).
It could get messy if we ever want to add some env vars to some user
containers but not others? But for now I don't think we should do that.