Made the pod parsing annotation errors easier to understand
Previously, we were not printing what the actual annotation
value was that we were erroring on! And sometimes not even the error!

This change makes us print out the annotation key, value and error
whenever we encounter a problem for maximum debugability.

Otherwise we get errors like

    * annotation is not a valid scheduler policy: pod.netflix.com/sched-policy

Which kinda makes you think that `*` was the value? But it wasn't.