Added annotation to store a pod termination reason
Pods can be deleted from a number of different parts of the codebase.
The happy path comes from the control-plane, but there are other paths.

This new proposed annotation will allow us to track the reason for the
pod termination.

Hopefully this annotation will reduce the need for having kubeapi audit
logs or similar for debugging what when wrong.

*And* it will also assist in giving pods metadata on shutdown about
"why" they are being shut down.